Update README.md
An Automobile company Axon, which is a retailer selling classic cars, is facing issues in managing and analyzing their sales data. The sales team is struggling to make sense of the data and they do not have a centralized system to manage and analyze the data. The management is unable to get accurate and up-to-date sales reports, which is affecting the decision-making process. For this,
I've Utilized SQL, PowerBI, Excel, and more to analyze sales, profit, orders, and identified most profitable countries and top customers. Overcame challenges and gained valuable insights.
This project provides valuable insights into the Axon Automobile sales data, revealing revenue trends, peak months, top customers, profitable products, and high-performing employees.